requirements.txt
streamlit
pandas
openpyxl
python-dotenv
litellm
google-adk